Fix OS lint failures for project create form
Rename CreateProjectSchema to CreateProjectInput and move
cacheCreatedProjectQueries into its own module so the component file only
exports components (satisfies zod-schema-naming and only-export-components).
